luarocks.bbclass: Fixup do_install to work with usrmerge feature.
When usrmerge is set as a distro feature, the "root_prefix" will
already be "/usr" and was resulting in components being installed
to "/usr/usr". Copying between the appropriate directories instead
resolves this.

Additionally, this change updates the copy logic to preserve only
links and not mode and permission to make the chown commands
unnecessary.

Signed-off-by: Charlie Johnston <charlie.johnston@loftorbital.com>